requirements.txt
streamlit>=1.30.0
pdfplumber>=0.10.0
ebooklib>=0.18
openai>=1.0.0
markdown>=3.5
python-dotenv>=1.0.0
